+++ src/instructions/ldi.rs
@@ -15,8 +15,8 @@ impl VM {
         // Use of wrapping_add for handling overflow cases (that happen when adding to a sign-extended negative number)
         let position = self.reg.pc.wrapping_add(pc_offset);
 
-        let temp_value = self.mem.read(position as usize);
-        let value = self.mem.read(temp_value as usize);
+        let temp_value = self.mem.read(position);
+        let value = self.mem.read(temp_value);
 
         self.reg.update(r0, value);
     }
