+++ src/instructions/ldr.rs
@@ -11,7 +11,7 @@ impl VM {
         let pc_offset = sign_extend(instr & 0b111111, 6);
 
         let final_address = self.reg.get(br).wrapping_add(pc_offset);
-        let value_read = self.mem.read(final_address as usize);
+        let value_read = self.mem.read(final_address);
         self.reg.update(dr, value_read);
     }
 }
